+++ CHANGELOG.md
@@ -5,8 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+- The bot will now crash if the commands have not been deployed.
+
+### Changed
+- `/help` now prints a clickable link to the `/suggest` command.
+
 ## [0.0.0] - 2023-07-02
 ### Added
 - Initial commit
 
+[Unreleased]: https://github.com/AverageHelper/Pippin/compare/v0.0.0...HEAD
 [0.0.0]: https://github.com/AverageHelper/Pippin/releases/tag/v0.0.0